% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_add_logical_significance.R
\name{tidy_lm_add_logical_p_values}
\alias{tidy_lm_add_logical_p_values}
\title{\code{tidy_lm_add_logical_significance}}
\usage{
tidy_lm_add_logical_p_values(tidy_df, p_values = p.value, alpha = 0.05,
  sig_column_name = "sig.05")
}
\arguments{
\item{tidy_df}{a \code{tidy} table of \code{lm} results}

\item{p_values}{= "p.value" (default). can specificy another column of p.values}

\item{alpha}{= 0.05 (default). can specify another alpha value}

\item{sig_column_name}{= "sig.05" (default). can specify another name for the new column (e.g., sig.alpha.value).}
}
\value{
logical column of significance (i.e., TRUE or FALSE) by comparing p
to the cut-off alpha. this column is added to a \code{tidy} table of a \code{lm}
}
\description{
add column of logical significance to lm tidy table
}
\examples{
data <- tibble(a = scale(sample.int(100), scale = F),
                           b = scale(sample.int(100), scale = F),
                           c = b^2,
                           d = scale(sample.int(100), scale = F))
lm(a ~ b, data) \%>\%
  tidy() \%>\%
  tidy_lm_add_logical_p_values()
}
\author{
Ekarin Eric Pongpipat
}
