% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_add_logical_significance.R
\name{tidy_lm_add_logical_significance}
\alias{tidy_lm_add_logical_significance}
\title{\code{tidy_lm_add_logical_significance}}
\usage{
tidy_lm_add_logical_significance(tidy_df, alpha = 0.05)
}
\arguments{
\item{tidy_df}{a \code{tidy} table of \code{lm} results}

\item{alpha}{= 0.05 (default). can specify another alpha value}

\item{p_values}{= "p.value" (default). can specificy another column of p.values}
}
\value{
logical column of significance (i.e., TRUE or FALSE) by comparing p
to the cut-off alpha. this column is added to a \code{tidy} table of a \code{lm} result
}
\description{
add a logical column of significance comparing p-values to alpha
}
\examples{
library(dplyr); library(broom); library(bootPerm)
model <- lm(salary ~ sex, carData::Salaries) \%>\%
  tidy() \%>\%
  tidy_lm_add_logical_significance(., c("0.01", "0.05"))
}
\author{
Ekarin Eric Pongpipat
}
