% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_add_multiple_comparison_correction.R
\name{tidy_lm_add_multiple_comparison_correction}
\alias{tidy_lm_add_multiple_comparison_correction}
\title{\code{tidy_lm_add_multiple_comparison_correction}}
\usage{
tidy_lm_add_multiple_comparison_correction(tidy_df, methods = "all")
}
\arguments{
\item{tidy_df}{a \code{tidy} table of \code{lm} results}

\item{methods}{= "all" (default). a string or vector \code{c()} of p.adjust methods,
which include: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", or "all"}
}
\value{
numeric adjusted p.values column using all or defined mutliple
comparison correction methods. this column is added to a \code{tidy} table
of a \code{lm} result
}
\description{
add adjusted p-values (i.e., multiple comparison correction methods)
as a new column using the \code{p.adjust()} function
}
\examples{
packages <- c("broom", "bootPermBroom", "dplyr")
xfun::pkg_attach(packages, message = F, install = T)

lm(salary ~ sex, carData::Salaries) \%>\%
  tidy() \%>\%
  tidy_lm_add_multiple_comparison_correction()
}
\author{
Ekarin Eric Pongpipat
}
