% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_bootstrap_permute.R
\name{tidy_lm_bootstrap_permute}
\alias{tidy_lm_bootstrap_permute}
\title{\code{tidy_lm_bootstrap_permute}}
\usage{
tidy_lm_bootstrap_permute(data, formula, n_bootstrap = 1000,
  bootstrap_ci = 0.95, n_permute = 1000, var_permute,
  r_squared_ci = 0.95)
}
\arguments{
\item{data}{a data.frame to be analyzed}

\item{formula}{a formula to be analyzed as typically written for the \code{lm} function}

\item{n_bootstrap}{= 1000 (default) the number of bootstraps to perform}

\item{bootstrap_ci}{= 0.95 (default)}

\item{n_permute}{= 1000 (default) the number of permutations to perform}

\item{var_permute}{variable(s) to unlink in the permutation}

\item{r_squared_ci}{= 0.95 (default)}
}
\value{
outputs \code{tidy} table that includes the p.value from the permutation of a \code{lm} test
}
\description{
\code{permute} \code{lm} and output the results as a \code{tidy} table.
}
\examples{
data <- tibble(
  a = scale(sample.int(100), scale = F),
  b = scale(sample.int(100), scale = F),
  c = b^2,
  d = scale(sample.int(100), scale = F)
)
tidy_lm_bootstrap_permute(data, a ~ b, var_permute = "a")
}
\author{
Ekarin Eric Pongpipat
}
