% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_permute.R
\name{tidy_lm_permute}
\alias{tidy_lm_permute}
\title{\code{permute_lm_tidy}}
\usage{
tidy_lm_permute(data, formula, n_permute = 1000, var_permute)
}
\arguments{
\item{data}{a data.frame to be analyzed}

\item{formula}{a formula to be analyzed as typically written for the \code{lm} function}

\item{n_permute}{= 1000 (default) the number of permutations to perform}

\item{var_permute}{variable(s) to unlink in the permutation}
}
\value{
outputs \code{tidy} table that includes the p.value from the permutation of a \code{lm} test
}
\description{
perform permutation testing and output permutation p-value in tidy format
}
\examples{
 set.seed(2019)
data <- tibble(a = scale(sample.int(100), scale = F),
              b = scale(sample.int(100), scale = F),
              c = b^2,
              d = scale(sample.int(100), scale = F))
df_tidy <- permute_lm_tidy(data = data, formula = "a ~ b + c + d", n_permute = 1000, var_permute = "a")
permute_lm_tidy(data = data, formula = "a ~ b + c", n_permute = 100, "a") \%>\% tidy_add_r_squared(., n = nrow(data))
}
\author{
Ekarin Eric Pongpipat
}
