% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm_add_logical_significance.R
\name{tidy_lm_add_logical_significance}
\alias{tidy_lm_add_logical_significance}
\title{\code{tidy_lm_add_logical_significance}}
\usage{
tidy_lm_add_logical_significance(tidy_df, p_values = "p.value",
  alpha = 0.05, sig_column_name = "sig.05")
}
\arguments{
\item{tidy_df}{a \code{tidy} table of \code{lm} results}

\item{p_values}{= "p.value" (default). can specificy another column of p.values}

\item{alpha}{= 0.05 (default). can specify another alpha value}

\item{sig_column_name}{= "sig.05" (default). can specify another name for the new column (e.g., sig.alpha.value).}
}
\value{
logical column of significance (i.e., TRUE or FALSE) by comparing p
to the cut-off alpha. this column is added to a \code{tidy} table of a \code{lm}
}
\description{
add a logical column of significance comparing p-values to alpha
}
\examples{
packages <- c("broom", "broomExtra", "dplyr", "modelr", "tibble")
xfun::pkg_attach2(packages, message = F)

data <- tibble(
  a = scale(sample.int(100), scale = F),
  b = scale(sample.int(100), scale = F),
  c = b^2,
  d = scale(sample.int(100), scale = F)
)

lm(a ~ b, data) \%>\%
  tidy() \%>\%
  tidy_lm_add_logical_significance()
}
\author{
Ekarin Eric Pongpipat
}
